Refactor datatypes into separate folder
