+++ Scripts/datatypes/path_analysis.py
@@ -1,0 +1,24 @@
+class PathAnalysis:
+    def __init__(self, link_component, od_values):
+        self.spec = {
+            "link_component": link_component,
+            "turn_component": None,
+            "operator": "+",
+            "selection_threshold": {
+                "lower": None,
+                "upper": None,
+            },
+            "path_to_od_composition": {
+                "considered_paths": "ALL",
+                "multiply_path_proportions_by": {
+                    "analyzed_demand": False,
+                    "path_value": True,
+                }
+            },
+            "analyzed_demand": None,
+            "results": {
+                "selected_link_volumes": None,
+                "selected_turn_volumes": None,
+                "od_values": od_values,
+            },
+        }
